feat(all): Initial implementation
This commit provides the initial implementation of the server. It is a
little more complex than it needs to be, with options defined but not
exposed over the terminal.

This is a kind of, one shot program I need now, and I will just improve
it as I need to.